refactor(commands): extract folders, documents, links domains (gated, verbatim)
Pure MOVE, zero behavior change. Three domains lift out of commands/mod.rs
into commands/{folders,documents,links}.rs (each #[path]-aliased to dodge E0255
vs crate::links / crate::storage::{links,folders_store}); mod.rs 11,447 -> 9,926
(-1,521). lib.rs generate_handler! BYTE-FOR-BYTE UNCHANGED.

18 #[tauri::command] moved (5 documents + 4 folders + 9 links); command count
conserved 92. Byte-identical (whole-region diffs vs trunk empty); only 3
fn->pub(crate) widenings, each forced by a retained lifecycle_tests call site
(insert_extracted_document, import_document_write_gate, build_folder_tree),
bodies unchanged.

Gate/guard logic UNCHANGED, only relocated:
- documents get_document keeps folder_is_unlocked gate; link_endpoint_is_unlocked
  keeps get_document_if_visible for the Document leg; list_documents / search
  stay visibility-gated.
- links list_link_candidates / resolve_wikilink keep visibility filtering verbatim
  (resolve_wikilink_hides_sealed_target passes).
- accept_link_inner moved VERBATIM: the scoped { } lifecycle guard released before
  the materialize loop is intact (the self-deadlock fix was NOT reintroduced).

Deliberately LEFT in mod.rs (shared with pipeline/notes/lock, per bias-to-move-less):
move_note + the whole move/auto-file/salvage seal cluster (move_into_locked_folder,
seal_moved_note, classify_auto_file_target, ...), the write-time link hooks
(index_wikilinks_best_effort, auto_link_semantic_best_effort), rederive_links_for_folder
(lock.rs), build_and_persist_entities + link_related_notes_inner (pipeline). The moved
commands reach them via super::.

Tests all stayed in mod.rs (#[path] test modules 8 = trunk). Verified: cargo build
--lib 0 warnings, full cargo test --lib 1992/0, targeted folder/document/link/accept/
visible/lifecycle 471/0 (incl resolve_wikilink_hides_sealed_target,
get_document_visibility_gated, sealed_document_excluded_from_search,
seal_preserves_dismissed_and_accepted_link_decisions), clippy --lib -D warnings clean,
MSRV grep clean.
