+++ src-tauri/src/commands/mod.rs
@@ -167,6 +167,50 @@ pub use lock_commands::*;
 mod org_commands;
 pub use org_commands::*;
 
+// DOCUMENT-INGEST command surface (upload/import/list/read/delete of brain `documents` — a GATED
+// domain: every write WRITE-GATES the folder + re-checks the gate before the plaintext INSERT, every
+// read masks a sealed-not-unlocked folder to empty/"", `delete_document` revokes org shares
+// before dropping the row; the gate/mask LOGIC is byte-identical, only relocated). Bound as
+// `documents_commands` (via `#[path]`) to keep it clearly distinct and avoid any future name shadow.
+// The glob re-export makes every moved command resolve UNCHANGED at `crate::commands::…` for
+// `generate_handler!` in `lib.rs` and every caller (incl. the STAYING test modules).
+#[path = "documents.rs"]
+mod documents_commands;
+pub use documents_commands::*;
+
+// FOLDERS command surface (create/list/rename/delete of meeting/note folders — a GATED domain:
+// `list_folders` folds the session unlock set into per-folder note counts, `delete_folder` refuses a
+// sealed-not-unlocked folder and PERMANENTLY unseals a session-unlocked one via `remove_lock_inner`
+// before demoting its notes to root, never orphaning encrypted content; the gate/seal LOGIC is
+// byte-identical, only relocated). The MEETING move-into-folder command (`move_note`) + its auto-file
+// / salvage seal helper web (`move_into_locked_folder`/`seal_moved_note`/`classify_auto_file_target`/
+// `seal_auto_filed_note`/`finalize_salvage_lock_state`/`AutoFileTarget`/`ensure_meeting_folder_target`)
+// deliberately STAY in `commands/mod.rs` — they are pipeline-called seal machinery, not folder CRUD.
+// Bound as `folders_commands` (via `#[path]`) to avoid colliding with the crate-level
+// `crate::storage::folders_store` and any future `folders` name (E0255). The glob re-export makes every
+// moved command resolve UNCHANGED at `crate::commands::…` for `generate_handler!` and every caller
+// (incl. the sibling `commands/notes.rs` → `rename_folder_inner`/`delete_folder_inner`).
+#[path = "folders.rs"]
+mod folders_commands;
+pub use folders_commands::*;
+
+// LINKS command surface (the note↔meeting↔document link engine surface — a GATED domain: `list_links`
+// gates BOTH endpoints, `accept_link`/`dismiss_link` gate + refuse behind a lock, `link_items`/
+// `unlink_items` gate both endpoints before any write, `list_link_candidates`/`resolve_wikilink` route
+// through the `*_visible` readers; the gate LOGIC is byte-identical, only relocated). `accept_link_inner`
+// keeps its EXACT scoped-lifecycle-guard body (the guard is scoped in a `{ }` block RELEASED before the
+// materialize loop so the composed `update_note_inner` can re-take the non-reentrant `Mutex<()>` without
+// self-DEADLOCK) — moved verbatim. The write-time index hooks (`index_wikilinks_best_effort`/
+// `auto_link_semantic_best_effort`), the unlock re-derive (`rederive_links_for_folder`), the entity
+// persist (`build_and_persist_entities`), and the `link_related_notes_inner` core (pipeline-called) STAY
+// in `commands/mod.rs` — the moved commands reach them through `use super::*` (a `commands` submodule sees
+// its parent's private items). Bound as `links_commands` (via `#[path]`) to avoid colliding with the
+// crate-level `crate::links` / `crate::storage::links` (E0255). The glob re-export makes every moved
+// command resolve UNCHANGED at `crate::commands::…` for `generate_handler!` and every caller.
+#[path = "links.rs"]
+mod links_commands;
+pub use links_commands::*;
+
 /// Keychain account for the AI Gateway API key (matches `summarize::GATEWAY_KEY_ACCOUNT`).
 /// Strictly separate from `ANTHROPIC_KEY_ACCOUNT` (defined in `commands/secrets.rs`) — never a
 /// fallback to the Anthropic key (R3). Kept here because the gateway model-listing helpers below
@@ -1865,15 +1909,6 @@ pub(crate) fn link_related_notes_inner(state: &AppState, meeting_id: &str) -> Re
     Ok(())
 }
 
-/// `link_related_notes(meeting_id)` — MANUAL re-link / backfill trigger for the Stage 2 / Lane A
-/// pass. The AUTO pipeline runs the same [`link_related_notes_inner`] as a deferred post-`Exported`
-/// pass; this command lets the user (or a backfill over old notes) re-run it on demand. Lock-gated +
-/// seal-safe (a sealed meeting is a silent no-op) and ZERO egress.
-#[tauri::command]
-pub fn link_related_notes(state: State<'_, AppState>, meeting_id: String) -> Result<(), AppError> {
-    link_related_notes_inner(state.inner(), &meeting_id)
-}
-
 /// brain2 realtime typed @brain notes — persist the user's free-text notes typed DURING a meeting
 /// (the FE autosaves the whole buffer here). The buffer (a) feeds the in-meeting brain's system
 /// prompt while recording and (b) is folded into the finalized note at summarize time.
@@ -2389,441 +2424,6 @@ pub(crate) fn get_fact_receipt_inner(
         "computed fact receipt"
     );
     Ok(receipt)
-}
-
-/// The extensions document ingestion accepts (Brain v3 PR-2). Text (md/txt) plus the extracted
-/// formats: PDF (macOS PDFKit, scanned-PDF pages fall back to on-device Vision OCR), DOCX/PPTX
-/// (pure-Rust OOXML), XLSX (calamine), HTML, and — Brain v3 OCR — direct image import
-/// (png/jpg/jpeg/heic/tiff/tif/bmp/gif) via on-device Apple Vision. Dispatch + extraction live in
-/// `crate::extract`; anything else is rejected with `InvalidArg`.
-const DOC_ALLOWED_EXTS: &[&str] = &[
-    "md", "txt", "pdf", "docx", "pptx", "xlsx", "html", "htm", "png", "jpg", "jpeg", "heic", "tiff",
-    "tif", "bmp", "gif",
-];
-
-/// Document ingestion — upload a local file INTO a folder so its EXTRACTED text is chunked + embedded
-/// into the on-device vector layer and the brain/Ask can retrieve it. Returns the new document id.
-/// ASYNC (Brain v3 PR-2): a large PDF's extract+chunk+embed runs off the UI thread behind the shared
-/// heavy-inference permit + the RAM floor, emitting counts-only progress via [`EVENT_DOC_IMPORT`].
-///
-/// LOCK-MODEL:
-/// - WRITE-GATE: refuse a sealed-and-NOT-session-unlocked folder (`AppError::Locked`) — an ungated
-///   write would land plaintext at rest behind the lock (mirrors `save_manual_notes`'s gate).
-/// - Extension allowlist — reject anything else with `AppError::InvalidArg`.
-/// - We store the EXTRACTED TEXT only (`documents.text`), never the source binary — no new seal path.
-/// - EMBED only when the REAL e5 model is present (`embed_model_present()`): otherwise the chunks are
-///   stored WITHOUT vectors (no stub vectors polluting the index — mirrors `should_auto_index`).
-/// - The text is SEALED-AND-RESTORED with the folder on lock/unlock; its chunks are PURGED on lock,
-///   re-embeddable on unlock.
-#[tauri::command]
-pub async fn import_document(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    path: String,
-    folder_id: String,
-) -> Result<String, AppError> {
-    // 0) WRITE-GATE up front, on the async task (touches the borrowed `&AppState`): a
-    //    sealed-and-NOT-session-unlocked folder is refused BEFORE any file work so the caller fails
-    //    fast (`AppError::Locked`), and an unknown folder is `InvalidArg`. The gate is RE-CHECKED
-    //    inside the blocking closure right before the plaintext INSERT (below) using the cloned
-    //    `unlocked_folders` handle, so a relock racing between here and the insert can't land
-    //    plaintext at rest behind the lock.
-    import_document_write_gate(state.inner(), &folder_id)?;
-
-    // Clone ONLY the `Arc` handles that cross into the blocking closure — `AppState` itself is not
-    // `Clone`. `db` for insert+index, `unlocked_folders` for the pre-insert gate re-check,
-    // `heavy_inference` for the ONE heavy permit. Everything past this point is `'static`.
-    let db = std::sync::Arc::clone(&state.inner().db);
-    let unlocked = std::sync::Arc::clone(&state.inner().unlocked_folders);
-    let sem = std::sync::Arc::clone(&state.inner().heavy_inference);
-
-    // RAM floor: under memory pressure, do the (still off-thread) extract+insert but SKIP the embed —
-    // the chunks + FTS are durable (keyword retrieval works) and the idempotent repair tick / a later
-    // Reindex fills the vectors. Never fail the import over a busy machine. Decided here (on the async
-    // task) and passed INTO the closure so the whole heavy pipeline stays inside ONE `run_heavy` scope.
-    let ram_permits = crate::transcribe::model::topic_backfill_ram_permits_now();
-
-    // The WHOLE pipeline — extract (whole-file read + zip/XML parse for OOXML/XLSX, or the per-page
-    // PDFKit loop; multi-second on a large doc), insert the row + chunk (FTS durable), then embed —
-    // runs OFF the Tokio worker behind the ONE heavy-inference permit (rust-tauri rule: long-running
-    // work never blocks the runtime thread). Progress events fire from inside across the stages
-    // (extracting → chunking → embedding → done); the `AppHandle` is `Clone` + `Send` so the emitter
-    // reaches the FE from the blocking thread. Best-effort per stage: the row/chunks stay durable even
-    // if the embed fails.
-    let app2 = app.clone();
-    let path2 = path.clone();
-    let folder2 = folder_id.clone();
-    crate::perf::run_heavy(&sem, move || {
-        crate::events::emit_doc_import(&app2, "", "extracting", 0, 0);
-        // EXTRACT (pure `path → text`, no DB/state). The progress closure translates the extractor's
-        // per-page signal into `EVENT_DOC_IMPORT` "extracting done/total" events (Fix 3: real counts)
-        // and records an OCR-cap truncation so the "done" event can flag a partial import (Fix 2).
-        let ocr_truncated = std::cell::Cell::new(false);
-        let (name, stored) = {
-            let app_p = &app2;
-            let extract_progress = |p: crate::extract::ExtractProgress| match p {
-                crate::extract::ExtractProgress::Page { done, total } => {
-                    crate::events::emit_doc_import(app_p, "", "extracting", done, total);
-                }
-                crate::extract::ExtractProgress::OcrTruncated { .. } => {
-                    ocr_truncated.set(true);
-                }
-            };
-            extract_document_text(&path2, &extract_progress)?
-        };
-
-        crate::events::emit_doc_import(&app2, "", "chunking", 0, 0);
-        // GATE (re-checked from the cloned handle) + INSERT + CHUNK-ONLY index (FTS durable now).
-        let id = insert_extracted_document(&db, &unlocked, &folder2, &name, &stored)?;
-
-        // EMBED only if the RAM floor permitted (else defer to the repair tick). Best-effort: a
-        // failure logs (no PII) and leaves the durable chunks/FTS + row in place. Per-sub-batch
-        // progress (Fix 3) streams "embedding done/total" as each embed sub-batch completes.
-        if ram_permits {
-            let embedder = crate::embed::embed_model_present().then(crate::embed::active_embedder);
-            crate::events::emit_doc_import(&app2, &id, "embedding", 0, 0);
-            let id_for_progress = id.clone();
-            let embed_progress = |done: usize, total: usize| {
-                crate::events::emit_doc_import(&app2, &id_for_progress, "embedding", done, total);
-            };
-            if let Err(e) =
-                index_document_row_kind_routed_progress(&db, &id, embedder.as_deref(), &embed_progress)
-            {
-                tracing::warn!(target: "rag", error = %e, document_id = %id, "import: embed failed (content stored)");
-            }
-        } else {
-            tracing::info!(target: "documents", document_id = %id, "import: RAM floor — deferring embed to repair tick");
-        }
-        // DONE — carry the OCR-cap truncation flag so the FE can surface a partial-import notice.
-        crate::events::emit_doc_import_done(&app2, &id, ocr_truncated.get());
-        Ok(id)
-    })
-    .await
-}
-
-/// Inner of [`import_document`] taking `&AppState` (so the gate + allowlist + EXTRACTION are
-/// unit-testable without a `tauri::State`). EXTRACTS the file to blocks, stores the serialized text,
-/// inserts the row (write-gated), and DOES NOT embed (the async wrapper embeds behind the RAM floor /
-/// permit; a unit test with the model absent still gets chunk-only indexing via `ingest_into_folder`).
-///
-/// This runs SYNCHRONOUSLY (the async command wrapper does the same extract→insert→embed pipeline
-/// entirely inside `perf::run_heavy` / `spawn_blocking` — see [`import_document`]); this seam exists
-/// only so the extract + gate + insert are testable without a `tauri::State` or a runtime. It has no
-/// production caller (the async wrapper composes `extract_document_text` + `insert_extracted_document`
-/// directly), so it is compiled only under `cfg(test)`.
-#[cfg(test)]
-pub(crate) fn import_document_inner(
-    state: &AppState,
-    path: &str,
-    folder_id: &str,
-) -> Result<String, AppError> {
-    // EXTRACT (allowlist + `path → text`, pure, no DB/state).
-    let (name, stored) = extract_document_text(path, &crate::extract::no_progress)?;
-    // GATE + INSERT + chunk-only index (the shared seam the async wrapper also uses).
-    insert_extracted_document(&state.db, &state.unlocked_folders, folder_id, &name, &stored)
-}
-
-/// EXTRACT a supported document to its storable text form — PURE `path → (display_name, text)`, with
-/// NO `AppState`/DB/keychain touch, so the whole (potentially multi-second: whole-file read + zip
-/// decompress + XML parse for OOXML/XLSX, or the per-page PDFKit loop) extraction can run OFF the
-/// Tokio runtime inside `run_heavy`. Fails CLOSED with `AppError::InvalidArg` for an unsupported /
-/// extension-less / unreadable / no-extractable-text file. Returns the file-name component as the
-/// display name (never an on-disk path — no PII in the stored name/logs).
-fn extract_document_text(
-    path: &str,
-    progress: &crate::extract::ProgressFn<'_>,
-) -> Result<(String, String), AppError> {
-    // Extension allowlist. Lowercased; an extension-less path is rejected.
-    let p = std::path::Path::new(path);
-    let ext = p
-        .extension()
-        .and_then(|e| e.to_str())
-        .map(|e| e.to_ascii_lowercase());
-    let ext = match ext {
-        Some(e) if DOC_ALLOWED_EXTS.contains(&e.as_str()) => e,
-        _ => {
-            return Err(AppError::InvalidArg(
-                "unsupported document type — import md, txt, pdf, docx, pptx, xlsx, html, or an image (png/jpg/heic/tiff/bmp/gif)".into(),
-            ))
-        }
-    };
-
-    // EXTRACT to blocks (page/heading preserved), then serialize to the storable text form. A
-    // non-UTF-8 / unreadable / malformed / scanned-PDF / zip-bomb / over-size file fails closed inside
-    // `extract_blocks` (the OOXML/XLSX decompression-ratio guard + the universal extracted-text
-    // ceiling live there — see extract/mod.rs + extract/ooxml.rs). `progress` streams per-page counts.
-    let blocks = crate::extract::extract_blocks(p, &ext, progress)?;
-    if blocks.is_empty() || blocks.iter().all(|b| b.text.trim().is_empty()) {
-        return Err(AppError::InvalidArg(
-            "this document has no extractable text".into(),
-        ));
-    }
-    let stored = crate::extract::blocks_to_stored_text(&blocks);
-
-    // The display name = the file name (component only — never an on-disk path with personal content
-    // in logs). Fallback to "document" if the path has no file-name component.
-    let name = p
-        .file_name()
-        .and_then(|n| n.to_str())
-        .map(|s| s.to_string())
-        .unwrap_or_else(|| "document".to_string());
-    Ok((name, stored))
-}
-
-/// The WRITE-GATE for an uploaded document, evaluated from `Arc` handles (not a `&AppState`), so it
-/// can be RE-CHECKED inside the blocking closure right before the plaintext INSERT — a relock racing
-/// between the up-front async gate and the insert can't land plaintext at rest behind a lock. Mirrors
-/// [`ingest_into_folder_opts`]'s gate exactly: unknown folder ⇒ `InvalidArg`; sealed-and-NOT-session-
-/// unlocked ⇒ `AppError::Locked`; open OR session-unlocked ⇒ Ok.
-fn insert_extracted_document(
-    db: &std::sync::Arc<crate::storage::Db>,
-    unlocked_folders: &std::sync::Arc<std::sync::Mutex<std::collections::HashSet<String>>>,
-    folder_id: &str,
-    name: &str,
-    stored: &str,
-) -> Result<String, AppError> {
-    // The folder must exist (so the FK holds + the gating has an anchor).
-    let folder = db
-        .folder_by_id(folder_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no folder {folder_id}")))?;
-
-    // WRITE-GATE: a sealed-and-not-session-unlocked folder is refused (never resurrect plaintext at
-    // rest behind a lock). Same predicate as `folder_is_unlocked`, evaluated from the cloned handle.
-    if folder.locked {
-        let session_unlocked = unlocked_folders
-            .lock()
-            .map_err(|_| AppError::Storage("unlocked-folders mutex poisoned".into()))?
-            .contains(folder_id);
-        if !session_unlocked {
-            return Err(AppError::Locked(
-                "this folder is locked — unlock it to add to the brain".into(),
-            ));
-        }
-    }
-
-    let id = uuid::Uuid::new_v4().to_string();
-    let created_at = chrono::Utc::now().timestamp_millis();
-    db.insert_document(&id, folder_id, name, stored, "document", created_at)?;
-
-    // CHUNK-ONLY inline (FTS durable now); the caller embeds under the RAM floor + heavy permit. A
-    // failure logs (no PII) and does NOT fail the ingest (the row + plaintext are durable; a later
-    // unlock re-chunk / reindex recovers the index).
-    if let Err(e) = index_document_row_kind_routed(db, &id, None) {
-        tracing::warn!(target: "rag", error = %e, "ingest: chunk failed (content stored)");
-    }
-
-    // PII rule: log only ids, the kind, and byte counts — never the text/name.
-    tracing::info!(
-        target: "documents",
-        document_id = %id,
-        folder_id = %folder_id,
-        kind = "document",
-        bytes = stored.len(),
-        "ingested document into brain"
-    );
-    Ok(id)
-}
-
-/// The WRITE-GATE for an uploaded document evaluated from a borrowed `&AppState` (the up-front,
-/// fail-fast check on the async task before any file work). Same predicate as
-/// [`insert_extracted_document`]'s re-check; delegates to the existing `folder_is_unlocked`.
-fn import_document_write_gate(state: &AppState, folder_id: &str) -> Result<(), AppError> {
-    let folder = state
-        .db
-        .folder_by_id(folder_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no folder {folder_id}")))?;
-    if folder.locked && !folder_is_unlocked(state, folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to add to the brain".into(),
-        ));
-    }
-    Ok(())
-}
-
-/// Ingest TYPED text as a brain `note` (the Brain page "+ Add note"). Same gated ingest path + seal
-/// + vector indexing as an uploaded document, just `kind="note"` and no file/extension step.
-#[tauri::command]
-pub fn import_text(
-    state: State<'_, AppState>,
-    name: String,
-    text: String,
-    folder_id: String,
-) -> Result<String, AppError> {
-    import_text_inner(state.inner(), &name, &text, &folder_id)
-}
-
-/// Inner of [`import_text`] taking `&AppState` (unit-testable gate). Empty text is refused.
-pub(crate) fn import_text_inner(
-    state: &AppState,
-    name: &str,
-    text: &str,
-    folder_id: &str,
-) -> Result<String, AppError> {
-    if text.trim().is_empty() {
-        return Err(AppError::InvalidArg("note text is empty".into()));
-    }
-    let name = if name.trim().is_empty() {
-        "note"
-    } else {
-        name.trim()
-    };
-    ingest_into_folder(state, folder_id, name, text, "note")
-}
-
-/// The SINGLE gated ingest path for a typed note (`kind="note"`): look up the folder, WRITE-GATE it
-/// (a sealed-not-unlocked folder is refused so content can never appear at rest behind a lock),
-/// insert the `documents` row, and index its chunks into the vector layer ONLY when the REAL e5 model
-/// is present (never stub vectors — mirrors `should_auto_index`). The row is sealed-and-restored +
-/// purged-on-lock identically to an uploaded document. Returns the new id.
-///
-/// An uploaded DOCUMENT (`kind="document"`) takes the sibling [`insert_extracted_document`] seam
-/// instead — same gate + insert + chunk-only index, but reachable from `Arc` handles so the whole
-/// extract→insert→embed pipeline runs off the Tokio runtime inside `run_heavy` (Brain v3 PR-2).
-fn ingest_into_folder(
-    state: &AppState,
-    folder_id: &str,
-    name: &str,
-    text: &str,
-    kind: &str,
-) -> Result<String, AppError> {
-    // The folder must exist (so the FK holds + the gating has an anchor).
-    let folder = state
-        .db
-        .folder_by_id(folder_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no folder {folder_id}")))?;
-
-    // WRITE-GATE: a sealed-and-not-session-unlocked folder is refused (never resurrect plaintext at
-    // rest behind a lock). One gate for every ingest path.
-    if folder.locked && !folder_is_unlocked(state, folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to add to the brain".into(),
-        ));
-    }
-
-    let id = uuid::Uuid::new_v4().to_string();
-    let created_at = chrono::Utc::now().timestamp_millis();
-    state
-        .db
-        .insert_document(&id, folder_id, name, text, kind, created_at)?;
-
-    // ALWAYS chunk (doc_chunks + the fts_doc_chunks triggers) so keyword retrieval works on a
-    // DEFAULT install — an ingested note must never be write-only memory. Vectors ONLY when the
-    // REAL e5 model is present (never write stub vectors). Best-effort: a failure logs (no PII) and
-    // does NOT fail the ingest (the row + plaintext are durable; a later unlock re-chunk / reindex
-    // recovers the index).
-    //
-    // KIND-ROUTED (PR-1 finding 4): a pasted note (`kind='note'`) can carry YAML front-matter in its
-    // raw `text`, which must NEVER be embedded/indexed (DESIGN §1a — tags/properties pollute the
-    // vectors + snippets). Route through the ONE front-matter-stripping seam so the ingest matches
-    // every other note-index path.
-    let embedder = crate::embed::embed_model_present().then(crate::embed::active_embedder);
-    if let Err(e) = index_document_row_kind_routed(&state.db, &id, embedder.as_deref()) {
-        tracing::warn!(target: "rag", error = %e, "ingest: chunk/embed failed (content stored)");
-    }
-
-    // PII rule: log only ids, the kind, and byte/char counts — never the text/name.
-    tracing::info!(
-        target: "documents",
-        document_id = %id,
-        folder_id = %folder_id,
-        kind = %kind,
-        bytes = text.len(),
-        "ingested into brain"
-    );
-    Ok(id)
-}
-
-/// List a folder's documents (metadata only — NO text). GATED: a sealed-and-NOT-session-unlocked
-/// folder returns an EMPTY list (masked — never surface even a document name behind the lock),
-/// exactly like the masked detail DTO drops note/segments.
-#[tauri::command]
-pub fn list_documents(
-    state: State<'_, AppState>,
-    folder_id: String,
-) -> Result<Vec<DocumentInfo>, AppError> {
-    list_documents_inner(state.inner(), &folder_id)
-}
-
-/// Inner of [`list_documents`] taking `&AppState` (unit-testable gate).
-pub(crate) fn list_documents_inner(
-    state: &AppState,
-    folder_id: &str,
-) -> Result<Vec<DocumentInfo>, AppError> {
-    if !folder_is_unlocked(state, folder_id)? {
-        return Ok(Vec::new()); // sealed-not-unlocked ⇒ masked (no ids, no names).
-    }
-    state.db.documents_in_folder(folder_id)
-}
-
-/// Read ONE document's full text. GATED: a sealed-and-NOT-session-unlocked folder returns "" (masked
-/// — never leak the document text), exactly like `get_manual_notes`.
-#[tauri::command]
-pub fn get_document(state: State<'_, AppState>, id: String) -> Result<String, AppError> {
-    get_document_inner(state.inner(), &id)
-}
-
-/// Inner of [`get_document`] taking `&AppState` (unit-testable gate).
-pub(crate) fn get_document_inner(state: &AppState, id: &str) -> Result<String, AppError> {
-    let Some((folder_id, _name, text)) = state.db.get_document(id)? else {
-        return Ok(String::new()); // unknown id → nothing.
-    };
-    if !folder_is_unlocked(state, &folder_id)? {
-        return Ok(String::new()); // sealed-not-unlocked ⇒ masked, never the stored text.
-    }
-    // Brain v3 PR-2: strip the block-structure markers a PR-2 upload stores in `text` — the FE gets
-    // clean readable text (a note / md / txt / legacy row has no markers → unchanged).
-    Ok(crate::extract::render_display_text(&text))
-}
-
-
-
-/// Permanently delete a document and cascade-delete its chunks + vectors. GATED: a
-/// sealed-and-NOT-session-unlocked folder is refused (`AppError::Locked`) so the lock state can't be
-/// mutated from behind the gate (consistent with `import_document`'s write-gate).
-///
-/// DELETE-CASCADE FIX (2026-07-15): `delete_document` is generic over BOTH `kind='document'`
-/// (imported/ingested files) and `kind='note'` rows (`brain.component.ts`'s `removeItem` can reach
-/// either), so it needs the SAME org-share revoke cascade as [`delete_note`] before dropping the local
-/// row — a `kind='note'` document deleted through THIS surface must not resurrect via the org feed
-/// either. `async` (was sync) because the revoke is a network round-trip.
-#[tauri::command]
-pub async fn delete_document(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-) -> Result<(), AppError> {
-    // Only a `kind='note'` row is ever tab-tracked (a plain ingested `kind='document'` has no tab —
-    // see `TabKind` in `tab-keys.ts`), so only fire the delete-fan-out event for that case: emitting
-    // it for an id nothing tracks is harmless, but this stays precise about what was actually deleted.
-    let was_note = matches!(state.db.get_note_row(&id), Ok(Some(_)));
-    delete_document_inner(state.inner(), &id).await?;
-    if was_note {
-        crate::events::emit_content_deleted(&app, "note", &id);
-    }
-    // The delete purged its audit findings (id-matched) — ping the FE inbox (count-only).
-    emit_audit_updated_after_purge(&app, state.inner());
-    Ok(())
-}
-
-/// Inner of [`delete_document`] taking `&AppState` (unit-testable gate). `async` for the org-share
-/// revoke cascade (network round-trip); the gate + DB delete themselves stay synchronous internally.
-pub(crate) async fn delete_document_inner(state: &AppState, id: &str) -> Result<(), AppError> {
-    let Some(folder_id) = state.db.folder_for_document(id)? else {
-        return Ok(()); // unknown id → idempotent no-op.
-    };
-    if !folder_is_unlocked(state, &folder_id)? {
-        return Err(AppError::Locked(
-            "this folder is locked — unlock it to delete a document".into(),
-        ));
-    }
-    // REVOKE-BEFORE-DELETE (Bug A root cause): tear down every LIVE org share of this exact source
-    // BEFORE the local row disappears, so the background org-sync tick can never re-pull a still-live
-    // server item back into the local replica after the user asked to delete it. Fails LOUD: a revoke
-    // failure (e.g. offline) aborts the delete rather than silently leaving a dangling live share.
-    revoke_org_shares_for_source(state, None, Some(id)).await?;
-    state.db.delete_document(id)?;
-    tracing::info!(target: "documents", document_id = %id, "document deleted");
-    Ok(())
 }
 
 // ── NOTES (authored `documents(kind='note')`) ────────────────────────────────────────────────────
@@ -5212,492 +4812,6 @@ fn file_stem_of(path: &str) -> String {
         .to_string()
 }
 
-
-/// Resolve the people + projects in a meeting note → persist them to the encrypted DB graph
-/// (always) and mirror them as `[[Person]]` / `[[Project]]` vault stubs (only when a vault is
-/// configured + the meeting's folder is unsealed). The graph self-assembles. The DB sink works
-/// even with no vault set — hence no hard "set a vault folder" error anymore.
-#[tauri::command]
-pub async fn link_meeting_entities(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    meeting_id: String,
-) -> Result<crate::summarize::graph::GraphPayload, AppError> {
-    // BLK-2b READ-GATE: a sealed-and-not-unlocked meeting's note is blanked; refuse to extract
-    // entities from it (would feed a cloud provider blank text + re-write vault stubs). Fail closed.
-    if !meeting_is_unlocked(state.inner(), &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to link entities".into(),
-        ));
-    }
-    let meeting = state
-        .db
-        .get_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no meeting with id {meeting_id}")))?;
-    let note = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg("this meeting has no note yet".into()))?;
-    let title = meeting
-        .title
-        .clone()
-        .unwrap_or_else(|| "Meeting".to_string());
-    build_and_persist_entities(&app, &state, &meeting_id, &title, &note.markdown).await
-}
-
-/// Parse an IPC link-endpoint kind string into [`crate::links::LinkKind`], or a clean `InvalidArg`.
-fn parse_link_kind(s: &str) -> Result<crate::links::LinkKind, AppError> {
-    crate::links::LinkKind::parse(s).ok_or_else(|| {
-        AppError::InvalidArg(format!(
-            "unknown link kind {s:?} (expected \"meeting\", \"note\", or \"document\")"
-        ))
-    })
-}
-
-/// Brain v3 PR-3 — every persisted link edge incident on `(kind, id)`, BOTH endpoints
-/// visibility-gated in [`crate::storage::db::Db::links_for_visible`] (the queried item must be
-/// visible or the list is empty — no existence leak — and each edge's neighbour is dropped unless it
-/// too is visible). Snapshots the LIVE session unlock set. `kind` is `"meeting" | "note" |
-/// "document"`. Dismissed edges are never returned; suggested (semantic) edges are, so the FE can
-/// render Accept/Dismiss.
-#[tauri::command]
-pub fn list_links(
-    state: State<'_, AppState>,
-    kind: String,
-    id: String,
-) -> Result<Vec<crate::storage::models::LinkEdge>, AppError> {
-    let link_kind = parse_link_kind(&kind)?;
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.links_for_visible(link_kind, &id, &unlocked)
-}
-
-/// Brain v3 PR-3 — ACCEPT a suggested (semantic) link: flip it `status='active'`,
-/// `created_by='accepted'`, AND materialize the neighbour's `[[Title]]` into whichever endpoint is a
-/// locally-owned, session-visible note (via the managed `apply_link_markers` block — the ONLY path
-/// that writes a semantic link into a `.md`; the auto pass never does).
-///
-/// GATED (PR-5): accept is ONLY for an unconfirmed SUGGESTION (`edge_type='semantic'`,
-/// `status='suggested'`) — anything else (a deterministic wikilink/companion, an already-active
-/// manual/accepted, or a dismissed tombstone) is refused `InvalidArg`. BOTH endpoints must be
-/// session-visible (`link_endpoint_is_unlocked` on each) — if EITHER is sealed-and-not-unlocked the
-/// accept is refused `AppError::Locked` (never activate an edge behind a lock, never reveal a locked
-/// neighbour by materializing a link to it). Idempotent (re-accepting an already-active edge is a
-/// no-op; the DB DO-UPDATE guard never downgrades it). The materialized marker is preserved across a
-/// later neighbour-seal and re-materialized on unlock (brain-v3 audit Fix 1/Fix 4).
-#[tauri::command]
-pub fn accept_link(state: State<'_, AppState>, id: i64) -> Result<(), AppError> {
-    accept_link_inner(state.inner(), id)
-}
-
-pub(crate) fn accept_link_inner(state: &AppState, id: i64) -> Result<(), AppError> {
-    // BLK-1 / TOCTOU + non-reentrant guard: SCOPE the lifecycle guard tightly around validate + gate
-    // + row-flip so a concurrent lock/relock cannot land between the endpoint gate and the flip. It is
-    // RELEASED before the note-body materialize below — `materialize_accepted_link` →
-    // `update_note_inner`/`update_note_doc_inner` take the guard THEMSELVES, so composing them under one
-    // held guard re-enters the non-reentrant `Mutex<()>` and self-DEADLOCKS on a valid accept (mirrors
-    // `link_items_inner`/`unlink_items_inner`). Lock order `lifecycle ⊃ db`.
-    let (src_kind, src_id, dst_kind, dst_id) = {
-        let _lifecycle = lifecycle_guard(state);
-        let Some((src_kind, src_id, dst_kind, dst_id, et, status)) = state.db.link_by_id(id)? else {
-            return Err(AppError::InvalidArg(format!("no link {id}")));
-        };
-        // ── Fix 5 (brain-v3 audit): accept is ONLY for an unconfirmed SUGGESTION. Refuse anything else
-        //    (a deterministic wikilink/companion, an already-active manual/accepted, or a dismissed
-        //    tombstone) — "accepting" those is meaningless and would let a caller flip arbitrary rows. ──
-        if !(et == "semantic" && status == "suggested") {
-            return Err(AppError::InvalidArg(format!(
-                "link {id} is not an acceptable suggestion (edge_type={et}, status={status})"
-            )));
-        }
-        // ── GATE both endpoints BEFORE flipping the row (never accept behind a lock, never reveal a
-        //    locked neighbour by activating an edge to it). Fail-closed on a sealed/unknown endpoint. ──
-        let (Some(sk), Some(dk)) = (
-            crate::links::LinkKind::parse(&src_kind),
-            crate::links::LinkKind::parse(&dst_kind),
-        ) else {
-            return Err(AppError::InvalidArg(format!("link {id} has a corrupt endpoint kind")));
-        };
-        if !link_endpoint_is_unlocked(state, sk, &src_id)?
-            || !link_endpoint_is_unlocked(state, dk, &dst_id)?
-        {
-            return Err(AppError::Locked(
-                "one of these items is locked — unlock it to accept the link".into(),
-            ));
-        }
-        // Flip the row first: the graph edge is active even if the .md materialize below is skipped
-        // (e.g. neither endpoint is a locally-owned editable note). The panel reflects the accept.
-        state.db.accept_link(id)?;
-        (src_kind, src_id, dst_kind, dst_id)
-    }; // ── lifecycle guard RELEASED here, before materialize (which re-takes it) ──
-    let unlocked = unlocked_snapshot(state)?;
-    // A semantic edge is undirected — materialize the neighbour's [[Title]] into whichever endpoint
-    // is a LOCAL, session-VISIBLE note we own the markdown of. Try src's note first, then dst's.
-    // Best-effort: a materialize skip/failure never rolls back the accept.
-    let endpoints = [
-        (src_kind.as_str(), src_id.as_str(), dst_kind.as_str(), dst_id.as_str()),
-        (dst_kind.as_str(), dst_id.as_str(), src_kind.as_str(), src_id.as_str()),
-    ];
-    for (owner_k, owner_id, other_k, other_id) in endpoints {
-        let (Some(owner_kind), Some(other_kind)) = (
-            crate::links::LinkKind::parse(owner_k),
-            crate::links::LinkKind::parse(other_k),
-        ) else {
-            continue;
-        };
-        // Resolve the neighbour's current gated title; skip if the neighbour is sealed (no leak).
-        let Some(title) =
-            state.db.link_endpoint_title_visible(other_kind, other_id, &unlocked)?
-        else {
-            continue;
-        };
-        if materialize_accepted_link(state, owner_kind, owner_id, &title)? {
-            break; // wrote (or found already-present) in one owned, visible source — done.
-        }
-    }
-    tracing::info!(target: "links", link_id = id, "accept_link");
-    Ok(())
-}
-
-/// Materialize the accepted neighbour's `[[Title]]` into the OWNER `(kind, id)`'s markdown via the
-/// managed `apply_link_markers` block — but ONLY when the owner is a LOCAL, session-VISIBLE note we
-/// own the markdown of (a meeting AI note, or an authored note). Returns `true` when the owner IS
-/// such a note (whether it wrote or the link was already present), `false` when the owner is not a
-/// writable/visible target (the caller then tries the other endpoint). Merges with the related-notes
-/// hits already rendered in the block so the auto block is preserved.
-fn materialize_accepted_link(
-    state: &AppState,
-    kind: crate::links::LinkKind,
-    id: &str,
-    title: &str,
-) -> Result<bool, AppError> {
-    let hit = crate::enrich::ContextHit {
-        source: match kind {
-            crate::links::LinkKind::Meeting => "meeting",
-            _ => "note",
-        }
-        .to_string(),
-        detail: format!("[[{title}]]"),
-        url: None,
-    };
-    match kind {
-        crate::links::LinkKind::Meeting => {
-            // GATE: the meeting's note must be session-visible to write plaintext.
-            if !meeting_is_unlocked(state, id)? {
-                return Ok(false);
-            }
-            let Some(existing) = state.db.get_latest_note_for_meeting(id)? else {
-                return Ok(false);
-            };
-            // Skip a sealed (blob-present) note (the seal-safety gate `link_related_notes_inner` uses).
-            let sealed = state
-                .db
-                .sealable_notes_for_meeting(id)?
-                .iter()
-                .any(|n| n.content_blob.is_some());
-            if sealed {
-                return Ok(false);
-            }
-            let merged = merge_related_hit(&existing.markdown, hit);
-            if merged != existing.markdown {
-                update_note_inner(state, id, &merged)?;
-            }
-            Ok(true)
-        }
-        crate::links::LinkKind::Note | crate::links::LinkKind::Document => {
-            let Some(row) = state.db.get_note_row(id)? else {
-                return Ok(false); // a raw document has no editable note markdown here.
-            };
-            if !folder_is_unlocked(state, &row.folder_id)? {
-                return Ok(false);
-            }
-            let title_disp = row
-                .title
-                .clone()
-                .filter(|t| !t.is_empty())
-                .unwrap_or_else(|| row.name.clone());
-            let merged = merge_related_hit(&row.text, hit);
-            if merged != row.text {
-                update_note_doc_inner(state, id, &title_disp, &merged)?;
-            }
-            Ok(true)
-        }
-    }
-}
-
-/// Append ONE `[[Title]]` [`ContextHit`] to a note's managed `murmur:links` block WITHOUT dropping
-/// any related-notes hits already rendered there. `apply_link_markers` is a full-block REPLACE, so we
-/// re-collect the existing rendered hits ([`crate::enrich::extract_link_hits`]), add the new one
-/// (deduped by rendered detail), and re-apply — the block stays idempotent and the auto related-notes
-/// entries survive an accept.
-fn merge_related_hit(markdown: &str, new_hit: crate::enrich::ContextHit) -> String {
-    let mut hits = crate::enrich::extract_link_hits(markdown);
-    if !hits.iter().any(|h| h.detail == new_hit.detail) {
-        hits.push(new_hit);
-    }
-    crate::enrich::apply_link_markers(markdown, &hits)
-}
-
-/// Brain v3 PR-3 — DISMISS a suggested link: TOMBSTONE it so no later auto pass re-suggests it. No
-/// markdown is touched (dismiss is graph-only). Idempotent.
-///
-/// Fix 5 (brain-v3 audit): dismissal is for SUGGESTIONS only, and is GATED. Refuse a dismiss on a
-/// DETERMINISTIC edge (`wikilink`/`companion`) — those are re-derived from the note body / companion
-/// link on every save, so a tombstone would be resurrected next save (a confusing no-op) AND could
-/// be abused to silently suppress a real structural link. A `manual` edge is removed via
-/// `unlink_items`, not dismissed. Both endpoints are gated (fail-closed) so a caller can neither
-/// dismiss behind a lock nor probe a locked neighbour's existence via the accept/refuse response.
-#[tauri::command]
-pub fn dismiss_link(state: State<'_, AppState>, id: i64) -> Result<(), AppError> {
-    dismiss_link_inner(state.inner(), id)
-}
-
-pub(crate) fn dismiss_link_inner(state: &AppState, id: i64) -> Result<(), AppError> {
-    let _lifecycle = lifecycle_guard(state);
-    let Some((src_kind, src_id, dst_kind, dst_id, et, _status)) = state.db.link_by_id(id)? else {
-        return Err(AppError::InvalidArg(format!("no link {id}")));
-    };
-    // Only a suggestion (semantic) or an accepted-then-regretted semantic edge is dismissable. A
-    // deterministic wikilink/companion edge is NOT (it would just come back); a manual edge is
-    // removed by `unlink_items`. Refuse the rest with a clear InvalidArg.
-    if et != "semantic" {
-        return Err(AppError::InvalidArg(format!(
-            "link {id} is a deterministic {et} edge — dismissal is for semantic suggestions (remove a manual link via unlink)"
-        )));
-    }
-    // ── GATE both endpoints (fail-closed on a sealed/unknown endpoint). ──
-    let (Some(sk), Some(dk)) = (
-        crate::links::LinkKind::parse(&src_kind),
-        crate::links::LinkKind::parse(&dst_kind),
-    ) else {
-        return Err(AppError::InvalidArg(format!("link {id} has a corrupt endpoint kind")));
-    };
-    if !link_endpoint_is_unlocked(state, sk, &src_id)?
-        || !link_endpoint_is_unlocked(state, dk, &dst_id)?
-    {
-        return Err(AppError::Locked(
-            "one of these items is locked — unlock it to dismiss the suggestion".into(),
-        ));
-    }
-    state.db.dismiss_link(id)?;
-    tracing::info!(target: "links", link_id = id, "dismiss_link");
-    Ok(())
-}
-
-/// note↔meeting-links PR-1 — is a link ENDPOINT `(kind, id)` session-VISIBLE right now? Mirrors
-/// [`materialize_accepted_link`]'s gate order: a `Meeting` endpoint gates on [`meeting_is_unlocked`];
-/// a `Note`/`Document` endpoint resolves its owning folder via `get_note_row` and gates on
-/// [`folder_is_unlocked`]. An UNKNOWN endpoint (no such note/document) reports `false` — fail-closed,
-/// there is nothing legitimate to link. Used by `link_items`/`unlink_items` to refuse `AppError::Locked`
-/// before any write, so a manual edge is never created behind a lock and never reveals a locked
-/// neighbour.
-fn link_endpoint_is_unlocked(
-    state: &AppState,
-    kind: crate::links::LinkKind,
-    id: &str,
-) -> Result<bool, AppError> {
-    match kind {
-        crate::links::LinkKind::Meeting => meeting_is_unlocked(state, id),
-        crate::links::LinkKind::Note => match state.db.get_note_row(id)? {
-            Some(row) => folder_is_unlocked(state, &row.folder_id),
-            None => Ok(false), // unknown note → nothing to surface. Fail-closed.
-        },
-        crate::links::LinkKind::Document => {
-            // An imported `document` (kind != 'note') is NOT a `get_note_row` row, so routing it
-            // through `get_note_row` refused it fail-closed EVEN WHEN VISIBLE (a spurious `Locked`
-            // on the +Link chooser). Gate it via the canonical visibility reader
-            // (`get_document_if_visible` applies `visibility_clause` against the live unlock set):
-            // `Some` ⇒ visible/unlocked, `None` ⇒ sealed-or-unknown ⇒ refuse. Documents stay
-            // linkable (the chooser AND the Ask source-picker both offer them) and still fail-closed.
-            let unlocked = unlocked_snapshot(state)?;
-            Ok(state.db.get_document_if_visible(id, &unlocked)?.is_some())
-        }
-    }
-}
-
-/// note↔meeting-links PR-1 — USER-INITIATED link: persist ONE directed `manual` edge
-/// `(src → dst)` AND, when the source is an OWNED note, materialize `[[dst Title]]` into its body.
-///
-/// GATE (BEFORE any write): BOTH endpoints must be session-VISIBLE — a `meeting` via
-/// [`meeting_is_unlocked`], a `note`/`document` via its folder ([`folder_is_unlocked`]). If either is
-/// sealed-and-not-session-unlocked → `AppError::Locked` (never link behind a lock, never reveal a
-/// locked neighbour). Unknown kinds are `AppError::InvalidArg`.
-///
-/// The `manual` row (`created_by='user'`, `status='active'`, `score=1.0`) is idempotent on the
-/// table's UNIQUE key. For a `note` SOURCE we own the markdown of, we ALSO materialize the neighbour's
-/// gated `[[Title]]` into the managed `murmur:links` block via the SAME [`materialize_accepted_link`]
-/// path the accept command uses — best-effort (a materialize skip never rolls back the row). A
-/// `meeting`/`document` source (no owned editable body) creates the row ONLY. The materialized
-/// wikilink is display-deduped against the manual edge in `links_for_visible`.
-#[tauri::command]
-pub fn link_items(
-    state: State<'_, AppState>,
-    src_kind: String,
-    src_id: String,
-    dst_kind: String,
-    dst_id: String,
-) -> Result<(), AppError> {
-    link_items_inner(state.inner(), &src_kind, &src_id, &dst_kind, &dst_id)
-}
-
-pub(crate) fn link_items_inner(
-    state: &AppState,
-    src_kind: &str,
-    src_id: &str,
-    dst_kind: &str,
-    dst_id: &str,
-) -> Result<(), AppError> {
-    let src = parse_link_kind(src_kind)?;
-    let dst = parse_link_kind(dst_kind)?;
-    // Refuse a self-link (a pair pointing at itself is meaningless and would pollute the graph).
-    if src == dst && src_id == dst_id {
-        return Err(AppError::InvalidArg("cannot link an item to itself".into()));
-    }
-    // BLK-1 / TOCTOU: SCOPE the lifecycle guard tightly around gate + row-write so a concurrent
-    // lock/relock cannot land between the visibility check and the edge upsert. It is RELEASED before
-    // the note-body materialize below — that callee (`update_note_doc_inner`) takes the guard ITSELF
-    // and re-checks the folder gate, so composing them under one held guard would re-enter a
-    // non-reentrant `Mutex<()>` (the lifecycle_guard doc: never hold it around a callee that takes
-    // it). Lock order `lifecycle ⊃ db`.
-    {
-        let _lifecycle = lifecycle_guard(state);
-        // ── GATE both endpoints BEFORE any write (fail-closed on a sealed/unknown endpoint). ──
-        if !link_endpoint_is_unlocked(state, src, src_id)?
-            || !link_endpoint_is_unlocked(state, dst, dst_id)?
-        {
-            return Err(AppError::Locked(
-                "one of these items is locked — unlock it to link".into(),
-            ));
-        }
-        // ── Persist the directed manual edge (idempotent on the UNIQUE key). ──
-        state
-            .db
-            .upsert_manual_link(src.as_str(), src_id, dst.as_str(), dst_id)?;
-    }
-    // ── Fork #2: a NOTE source ALSO gets the neighbour's [[Title]] in its body (best-effort). ──
-    // A meeting/document source has no owned editable markdown → row only. `materialize_accepted_link`
-    // re-gates the source folder itself before writing plaintext (never behind a lock).
-    if matches!(src, crate::links::LinkKind::Note) {
-        let unlocked = unlocked_snapshot(state)?;
-        // Resolve the dst's CURRENT gated title (skip if sealed — no leak; the gate above already
-        // proved it visible, so this normally resolves).
-        if let Some(title) = state.db.link_endpoint_title_visible(dst, dst_id, &unlocked)? {
-            // Reuse the EXACT accept-materialize path (managed `murmur:links` block, merge-preserving).
-            if let Err(e) = materialize_accepted_link(state, src, src_id, &title) {
-                // Best-effort: the graph row is authoritative; a markdown-write failure never fails the
-                // link (no PII — ids/error only).
-                tracing::warn!(
-                    target: "links",
-                    error = %e,
-                    "manual link materialize skipped (row persisted)"
-                );
-            }
-        }
-    }
-    tracing::info!(
-        target: "links",
-        src_kind = src.as_str(),
-        dst_kind = dst.as_str(),
-        "link_items"
-    );
-    Ok(())
-}
-
-/// note↔meeting-links PR-1 — REMOVE a user-initiated link: delete ONLY the directed `manual` edge
-/// `(src → dst)` and, when the source is an OWNED note, strip the matching `[[dst Title]]` from its
-/// managed `murmur:links` block. NEVER touches a `wikilink`/`companion`/`semantic` row for the pair.
-///
-/// GATE: BOTH endpoints must be session-VISIBLE (same gate as `link_items`) — never mutate a note's
-/// body behind a lock, never reveal a locked neighbour's title. The strip is BEST-EFFORT (a failure
-/// logs and never fails the unlink — the graph row is the authoritative removal). Unknown kinds are
-/// `AppError::InvalidArg`.
-#[tauri::command]
-pub fn unlink_items(
-    state: State<'_, AppState>,
-    src_kind: String,
-    src_id: String,
-    dst_kind: String,
-    dst_id: String,
-) -> Result<(), AppError> {
-    unlink_items_inner(state.inner(), &src_kind, &src_id, &dst_kind, &dst_id)
-}
-
-pub(crate) fn unlink_items_inner(
-    state: &AppState,
-    src_kind: &str,
-    src_id: &str,
-    dst_kind: &str,
-    dst_id: &str,
-) -> Result<(), AppError> {
-    let src = parse_link_kind(src_kind)?;
-    let dst = parse_link_kind(dst_kind)?;
-    // SCOPE the guard around gate + row-delete only; release before the note-body strip below (which
-    // re-enters the guard via `update_note_doc_inner`). Same non-reentrancy discipline as `link_items`.
-    {
-        let _lifecycle = lifecycle_guard(state);
-        // ── GATE both endpoints (never mutate a note body / reveal a neighbour behind a lock). ──
-        if !link_endpoint_is_unlocked(state, src, src_id)?
-            || !link_endpoint_is_unlocked(state, dst, dst_id)?
-        {
-            return Err(AppError::Locked(
-                "one of these items is locked — unlock it to unlink".into(),
-            ));
-        }
-        // ── Delete ONLY the manual edge (wikilink/companion/semantic rows for the pair untouched). ──
-        state
-            .db
-            .delete_manual_link(src.as_str(), src_id, dst.as_str(), dst_id)?;
-    }
-    // ── A NOTE source: strip the matching [[Title]] from its managed block (best-effort). ──
-    if matches!(src, crate::links::LinkKind::Note) {
-        let unlocked = unlocked_snapshot(state)?;
-        if let Some(title) = state.db.link_endpoint_title_visible(dst, dst_id, &unlocked)? {
-            if let Err(e) = strip_manual_link_marker(state, src_id, &title) {
-                tracing::warn!(
-                    target: "links",
-                    error = %e,
-                    "manual link marker strip skipped (row removed)"
-                );
-            }
-        }
-    }
-    tracing::info!(
-        target: "links",
-        src_kind = src.as_str(),
-        dst_kind = dst.as_str(),
-        "unlink_items"
-    );
-    Ok(())
-}
-
-/// note↔meeting-links PR-1 — remove ONE `[[title]]` [`ContextHit`] from an owned note's managed
-/// `murmur:links` block, re-applying the block with that hit filtered out (the INVERSE of
-/// [`merge_related_hit`]). Reuses [`crate::enrich::extract_link_hits`] +
-/// [`crate::enrich::apply_link_markers`] so the block stays idempotent and any auto related-notes /
-/// accepted-semantic hits that lived alongside it survive. WRITE-GATED via `update_note_doc_inner`
-/// (refuses a sealed folder). A no-op (the note is unchanged) when the block never carried the hit.
-fn strip_manual_link_marker(state: &AppState, note_id: &str, title: &str) -> Result<(), AppError> {
-    let Some(row) = state.db.get_note_row(note_id)? else {
-        return Ok(()); // no owned note markdown → nothing to strip.
-    };
-    let marker = format!("[[{title}]]");
-    let mut hits = crate::enrich::extract_link_hits(&row.text);
-    let before = hits.len();
-    hits.retain(|h| h.detail != marker);
-    if hits.len() == before {
-        return Ok(()); // the marker was not in the managed block → note unchanged.
-    }
-    let merged = crate::enrich::apply_link_markers(&row.text, &hits);
-    if merged != row.text {
-        let title_disp = row
-            .title
-            .clone()
-            .filter(|t| !t.is_empty())
-            .unwrap_or_else(|| row.name.clone());
-        update_note_doc_inner(state, note_id, &title_disp, &merged)?;
-    }
-    Ok(())
-}
-
 /// Brain v3 PR-3 — WRITE-TIME wikilink indexing hook, called from every note-save funnel AFTER the
 /// text is durable. Resolves `[[Title]]` → TARGET IDS and (delete-then-insert) stores this source's
 /// `wikilink` edges. BEST-EFFORT: a failure logs (ids/counts only, no PII) and never fails the save
@@ -5842,84 +4956,6 @@ pub(crate) fn rederive_links_for_folder(
         }
         Err(e) => tracing::warn!(target: "links", error = %e, "unlock companion re-derive: meeting-id list failed"),
     }
-}
-
-/// Resolve a clicked `[[Title]]` wikilink to a VISIBLE note/meeting/org-item to navigate to.
-/// Returns `None` when nothing matches OR the only local match is a
-/// sealed-and-not-session-unlocked target (gated in `Db::resolve_wikilink`) — so a wikilink click
-/// never reveals or opens locked content. The org leg (2026-07-15) is gated by membership +
-/// per-instance-enabled, the same seam `list_link_candidates`/`search_org_brain_hits` use. The FE
-/// routes on `kind` (`"meeting"` | `"note"` | `"org"`), or offers to create a note when `None`.
-#[tauri::command]
-pub fn resolve_wikilink(
-    state: State<'_, AppState>,
-    title: String,
-) -> Result<Option<crate::storage::models::WikiTarget>, AppError> {
-    let unlocked = unlocked_snapshot(state.inner())?;
-    state.db.resolve_wikilink(&title, &unlocked)
-}
-
-/// Live keystroke-prefix candidates for the inline `[[` / slash-menu link-insertion autocomplete
-/// (note-editor Fix 2). Distinct from [`resolve_wikilink`] (exact-title resolve on Enter/click) and
-/// from `note_assistant_action`'s `find_related` (SELECTION+semantic retrieval — the wrong shape
-/// for filtering on a short, growing keystroke prefix): this is a lightweight, gated title-prefix
-/// scan. GATED exactly like every other content read: notes/meetings go through
-/// `Db::list_link_candidates_visible` (`visibility_clause` on both legs, same as `resolve_wikilink`);
-/// org items go through `search_org_brain_hits`, the SAME retrieval-only, membership+enabled-gated,
-/// zero-egress reader `find_related` already uses (never a provider/egress call). Reuses
-/// [`crate::storage::models::NoteCitation`] — the popover renders it exactly like a `find_related`
-/// citation row, and `kind == "org"` carries an org item id (never a local id), matching that
-/// contract verbatim.
-///
-/// PAGINATED (2026-07-17 — the picker is an infinite scroll over the whole vault now, not a fixed
-/// top-8): one call returns the `limit`-sized page at `offset` of the stable combined ordering
-/// [visible notes] ++ [visible meetings] ++ [org hits] (org only for a non-empty prefix, folded in
-/// after the local total the Db reader reports). The FE owns its page size; the clamp keeps one
-/// IPC reply from ever dumping an unbounded slice of the vault.
-#[tauri::command]
-pub fn list_link_candidates(
-    state: State<'_, AppState>,
-    prefix: String,
-    offset: Option<u32>,
-    limit: Option<u32>,
-) -> Result<Vec<crate::storage::models::NoteCitation>, AppError> {
-    const DEFAULT_PAGE: i64 = 40;
-    const MAX_PAGE: i64 = 100;
-    let limit = limit.map_or(DEFAULT_PAGE, i64::from).clamp(1, MAX_PAGE);
-    let offset = i64::from(offset.unwrap_or(0));
-    let unlocked = unlocked_snapshot(state.inner())?;
-    let (mut out, local_total) =
-        state
-            .db
-            .list_link_candidates_visible(&prefix, limit, offset, &unlocked)?;
-    if (out.len() as i64) < limit {
-        let config = {
-            state
-                .config
-                .lock()
-                .map_err(|_| AppError::Config("config mutex poisoned".into()))?
-                .clone()
-        };
-        let q = prefix.trim();
-        if !q.is_empty() && crate::tools::org_brain_available(&state.db, &config) {
-            // Earlier pages consumed `local_total` local rows, then `offset - local_total`
-            // org rows once the offset ran past the local legs — skip exactly those. The
-            // org reader is bounded (≤20 per leg pre-fusion), so skip/take over its Vec
-            // is real pagination, not a hidden re-scan.
-            let org_skip = (offset - local_total).max(0) as usize;
-            let remaining = (limit - out.len() as i64).max(0) as usize;
-            let org_hits = crate::tools::search_org_brain_hits(&state.db, &config, q)?;
-            for hit in org_hits.into_iter().skip(org_skip).take(remaining) {
-                out.push(crate::storage::models::NoteCitation {
-                    kind: "org".into(),
-                    id: hit.item_id,
-                    title: hit.title,
-                    snippet: hit.snippet,
-                });
-            }
-        }
-    }
-    Ok(out)
 }
 
 /// Inner of [`get_person_dossier`] taking `&AppState` (unit-testable gate). `None` — an unknown id
@@ -8242,114 +7278,6 @@ pub fn toggle_bar(app: AppHandle) {
 // KEK-wrapped CK in `folders.wrapped_key`. Session-unlock decrypts back into the markdown column
 // for the session (no re-export). relock re-blanks. remove_lock is permanent (re-exports).
 
-/// Build the folder tree (roots → children) from the flat folder list + per-folder note counts +
-/// the current session unlock set.
-#[tauri::command]
-pub fn list_folders(state: State<'_, AppState>) -> Result<Vec<FolderNode>, AppError> {
-    let folders = state.db.list_folders()?;
-    let unlocked = {
-        state
-            .unlocked_folders
-            .lock()
-            .map_err(|_| AppError::Storage("unlocked-folders mutex poisoned".into()))?
-            .clone()
-    };
-    // Gated by the session unlock set — a sealed-and-not-unlocked folder's notes must not
-    // contribute to note_count (see count_notes_per_folder doc + .claude/rules/lock-model.md).
-    let counts = state.db.count_notes_per_folder(&unlocked)?;
-    let kinds = state.db.folder_kinds()?;
-    Ok(build_folder_tree(&folders, &counts, &unlocked, &kinds))
-}
-
-/// Assemble `FolderNode` roots (parent_id == None) and recurse children. Sealed-but-session-
-/// unlocked folders carry `unlocked = true`.
-fn build_folder_tree(
-    folders: &[Folder],
-    counts: &std::collections::HashMap<String, usize>,
-    unlocked: &std::collections::HashSet<String>,
-    kinds: &std::collections::HashMap<String, String>,
-) -> Vec<FolderNode> {
-    fn node(
-        f: &Folder,
-        folders: &[Folder],
-        counts: &std::collections::HashMap<String, usize>,
-        unlocked: &std::collections::HashSet<String>,
-        kinds: &std::collections::HashMap<String, String>,
-    ) -> FolderNode {
-        let children = folders
-            .iter()
-            .filter(|c| c.parent_id.as_deref() == Some(f.id.as_str()))
-            .map(|c| node(c, folders, counts, unlocked, kinds))
-            .collect();
-        FolderNode {
-            id: f.id.clone(),
-            name: f.name.clone(),
-            parent_id: f.parent_id.clone(),
-            note_count: counts.get(&f.id).copied().unwrap_or(0),
-            locked: f.locked,
-            unlocked: f.locked && unlocked.contains(&f.id),
-            kind: kinds
-                .get(&f.id)
-                .cloned()
-                .unwrap_or_else(|| "meeting".to_string()),
-            children,
-        }
-    }
-    folders
-        .iter()
-        .filter(|f| f.parent_id.is_none())
-        .map(|f| node(f, folders, counts, unlocked, kinds))
-        .collect()
-}
-
-/// Create a folder under an optional parent. The vault-relative path is derived from the parent
-/// path + the sanitized folder name; the matching vault subdirectory is created on disk.
-#[tauri::command]
-pub fn create_folder(
-    state: State<'_, AppState>,
-    name: String,
-    parent_id: Option<String>,
-) -> Result<Folder, AppError> {
-    let clean = crate::summarize::organize::sanitize_folder(&name)
-        .ok_or_else(|| AppError::InvalidArg("folder name is empty or invalid".into()))?;
-
-    // Resolve the parent's vault-relative path (if any) and compose the child path.
-    let parent_path = match parent_id.as_deref() {
-        Some(pid) => {
-            let parent = state
-                .db
-                .folder_by_id(pid)?
-                .ok_or_else(|| AppError::InvalidArg(format!("no parent folder {pid}")))?;
-            Some(parent.path)
-        }
-        None => None,
-    };
-    let rel_path = match &parent_path {
-        Some(p) if !p.is_empty() => format!("{p}/{clean}"),
-        _ => clean.clone(),
-    };
-
-    // Create the vault subdirectory (best-effort but surfaced): only when a vault is configured.
-    // D5: canonicalize + assert the composed dir stays inside the vault root before any mkdir.
-    if let Some(vault) = vault_path(&state) {
-        let vault_root = std::path::Path::new(&vault);
-        let dir = assert_in_vault(vault_root, std::path::Path::new(&rel_path))?;
-        std::fs::create_dir_all(&dir)
-            .map_err(|e| AppError::Export(format!("create folder dir failed: {e}")))?;
-    }
-
-    let folder = Folder {
-        id: uuid::Uuid::new_v4().to_string(),
-        name: clean,
-        path: rel_path,
-        parent_id,
-        locked: false,
-        created_at: chrono::Utc::now().to_rfc3339(),
-    };
-    state.db.insert_folder(&folder)?;
-    Ok(folder)
-}
-
 /// Reject filing a MEETING under a NOTE folder — the folder namespaces are disjoint (a meeting's
 /// `.md` lives under a meeting folder; a note under a note folder). Called before any DB reassign
 /// or FS move so a cross-namespace target can never take effect. `None` (the vault root) is always
@@ -8952,455 +7880,6 @@ pub(crate) fn stop_all_capture(state: &AppState) {
         .lock()
         .ok()
         .and_then(|mut slot| slot.take());
-}
-
-/// Rename a folder: change its display `name` (and the matching vault subdirectory + every governed
-/// `path`) without ever touching sealed content.
-///
-/// Steps, ordered so a crash never loses content:
-///  1. Sanitize the new name (same component-safe rule as `create_folder`; reject `/`, `..`, NUL).
-///  2. Recompose this folder's vault-relative path = parent path + sanitized name.
-///  3. If a vault is configured, MOVE the on-disk subdir `old_path` → `new_path` (best-effort rename;
-///     a missing source is fine). The dir holds only the OPEN folder's plaintext `.md`s — sealed
-///     folders keep their `.md`s deleted, so a locked-folder rename just renames an empty/absent dir.
-///  4. Update the `folders` row (name + path) and re-prefix the path of EVERY descendant folder, and
-///     re-point EVERY affected note's `exported_path` from `old_path/...` → `new_path/...`. Sealed
-///     notes have `exported_path = NULL` and are skipped — a LOCKED folder rename is metadata-only and
-///     never reaches the sealed blob / wrapped key (no decrypt, no re-seal).
-///
-/// Idempotent-ish: renaming to the same (sanitized) name is a no-op move + a column rewrite to the
-/// same values.
-#[tauri::command]
-pub fn rename_folder(
-    state: State<'_, AppState>,
-    folder_id: String,
-    new_name: String,
-) -> Result<Folder, AppError> {
-    rename_folder_inner(state.inner(), folder_id, new_name)
-}
-
-/// Inner of [`rename_folder`] taking `&AppState` (so tests can drive it without a `tauri::State`).
-/// Holds the [`AppState::lifecycle`] guard across the whole rename (path rewrites the seal/unseal
-/// lifecycle keys FS ops off — see the command doc).
-pub(crate) fn rename_folder_inner(
-    state: &AppState,
-    folder_id: String,
-    new_name: String,
-) -> Result<Folder, AppError> {
-    // BLK-1: serialize with the rest of the lock state machine. A rename never decrypts, but it
-    // rewrites `path` columns that the seal/unseal lifecycle keys vault FS ops off — hold the guard
-    // so it can't interleave with a concurrent lock/unlock/remove that also rewrites paths.
-    let _lifecycle = lifecycle_guard(state);
-
-    let clean = crate::summarize::organize::sanitize_folder(&new_name)
-        .ok_or_else(|| AppError::InvalidArg("folder name is empty or invalid".into()))?;
-
-    let folder = state
-        .db
-        .folder_by_id(&folder_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no folder {folder_id}")))?;
-    let old_path = folder.path.clone();
-
-    // Recompose this folder's path from its PARENT's path + the new sanitized name.
-    let parent_path = match folder.parent_id.as_deref() {
-        Some(pid) => state.db.folder_by_id(pid)?.map(|p| p.path),
-        None => None,
-    };
-    let new_path = match parent_path.as_deref() {
-        Some(p) if !p.is_empty() => format!("{p}/{clean}"),
-        _ => clean.clone(),
-    };
-
-    // No-op fast path: same path AND same name → nothing to move/rewrite.
-    if new_path == old_path && clean == folder.name {
-        return Ok(Folder {
-            name: clean,
-            path: new_path,
-            ..folder
-        });
-    }
-
-    // Move the on-disk vault subdir, if a vault is configured. Both ends are containment-checked.
-    // `std::fs::rename` moves the WHOLE subtree (including descendant `.md`s) in one atomic op.
-    let mut vault_configured = false;
-    if new_path != old_path {
-        if let Some(vault) = vault_path(state) {
-            vault_configured = true;
-            let vault_root = std::path::Path::new(&vault);
-            // Destination must stay inside the vault; the source is an existing in-vault dir.
-            let dest = assert_in_vault(vault_root, std::path::Path::new(&new_path))?;
-            let src = assert_in_vault(vault_root, std::path::Path::new(&old_path))?;
-            if src.exists() {
-                if let Some(parent) = dest.parent() {
-                    std::fs::create_dir_all(parent).map_err(|e| {
-                        AppError::Export(format!("create rename parent dir failed: {e}"))
-                    })?;
-                }
-                // A plain rename within the same vault is atomic on the same filesystem.
-                std::fs::rename(&src, &dest)
-                    .map_err(|e| AppError::Export(format!("rename folder dir failed: {e}")))?;
-            } else {
-                // Source absent (a locked folder's dir, or never materialized): ensure the
-                // destination exists so future plaintext `.md`s land in the renamed dir.
-                std::fs::create_dir_all(&dest).map_err(|e| {
-                    AppError::Export(format!("create renamed folder dir failed: {e}"))
-                })?;
-            }
-        }
-    }
-
-    // Rewrite the DB: this folder's name+path, then re-prefix every DESCENDANT folder's path. Order
-    // doesn't risk content loss — no markdown/blob column is touched; only path strings move.
-    state.db.rename_folder(&folder_id, &clean, &new_path)?;
-    if new_path != old_path {
-        reprefix_descendant_folder_paths(state, &folder_id, &new_path)?;
-        // Re-derive every governed note's `exported_path` to point under its (possibly renamed)
-        // folder's NEW on-disk dir. We rebuild from the file basename + the folder's new dir rather
-        // than swapping path prefixes — robust to `/var` vs `/private/var` canonicalization drift in
-        // the stored absolute path. The `fs::rename` already moved the bytes; this only re-points the
-        // DB. Sealed notes (NULL exported_path) are skipped. Walks this folder + the whole subtree.
-        if vault_configured {
-            reexport_notes_under_subtree(state, &folder_id)?;
-        }
-    }
-
-    Ok(Folder {
-        name: clean,
-        path: new_path,
-        ..folder
-    })
-}
-
-/// Recursively re-prefix the vault-relative `path` of every DESCENDANT folder of `folder_id` to sit
-/// under `new_prefix` after the folder itself was renamed. Walks the tree one level at a time via
-/// [`Db::child_folders`]; each child's recomposed path is `new_prefix` + the child's own name (so the
-/// rewrite is structural, not a brittle string-replace). Does NOT touch the child's `name`, lock
-/// state, or any note content — only the `path` column (the descendants' notes are re-pointed by the
-/// single absolute-dir swap in the caller, since `fs::rename` moved the whole subtree at once).
-fn reprefix_descendant_folder_paths(
-    state: &AppState,
-    folder_id: &str,
-    new_prefix: &str,
-) -> Result<(), AppError> {
-    for child in state.db.child_folders(folder_id)? {
-        let child_old = child.path.clone();
-        let child_new = if new_prefix.is_empty() {
-            child.name.clone()
-        } else {
-            format!("{new_prefix}/{}", child.name)
-        };
-        if child_new != child_old {
-            state.db.rename_folder(&child.id, &child.name, &child_new)?;
-        }
-        // Recurse into this child's own subtree.
-        reprefix_descendant_folder_paths(state, &child.id, &child_new)?;
-    }
-    Ok(())
-}
-
-/// After a folder rename moved the on-disk subtree, re-point the `exported_path` of every governed
-/// note in `folder_id` AND its descendants to its folder's NEW vault dir. Each note's new path is
-/// `<vault>/<folder.path>/<basename of the old exported_path>` (the `fs::rename` preserved the
-/// filename). Rebuilding from the basename (not a string-prefix swap on the stored absolute path) is
-/// robust to canonicalization drift (`/var` vs `/private/var`) and to where the original export wrote
-/// the path. Sealed notes carry `exported_path = NULL` and are skipped. Requires a configured vault.
-fn reexport_notes_under_subtree(state: &AppState, folder_id: &str) -> Result<(), AppError> {
-    let Some(vault) = vault_path(state) else {
-        return Ok(());
-    };
-    let vault_root = std::path::Path::new(&vault);
-
-    let folder = match state.db.folder_by_id(folder_id)? {
-        Some(f) => f,
-        None => return Ok(()),
-    };
-    // The folder's NEW absolute dir (containment-checked).
-    let new_dir = assert_in_vault(vault_root, std::path::Path::new(&folder.path))?;
-
-    for n in state.db.notes_in_folder(folder_id)? {
-        let Some(old) = n.exported_path else {
-            continue; // sealed note (no .md) — nothing to re-point.
-        };
-        let Some(name) = std::path::Path::new(&old).file_name() else {
-            continue;
-        };
-        let new_path = new_dir.join(name);
-        state.db.set_note_exported_path(
-            &n.meeting_id,
-            &n.provider_id,
-            &new_path.to_string_lossy(),
-        )?;
-    }
-
-    // Recurse into descendant folders (their dirs moved with the same single `fs::rename`).
-    for child in state.db.child_folders(folder_id)? {
-        reexport_notes_under_subtree(state, &child.id)?;
-    }
-    Ok(())
-}
-
-/// Delete a folder, NEVER losing a note. SECURITY-CRITICAL — a folder may hold notes and may be
-/// sealed (LOCKED). Rules, fail-closed:
-///
-///  - **Has child folders →** REJECT (`InvalidArg`). The FE deletes leaf-first; refusing here keeps
-///    a subtree from being silently orphaned (a child's `parent_id` would dangle).
-///  - **LOCKED + NOT session-unlocked →** REJECT (`AppError::Locked`). We have no CK to unseal the
-///    folder's notes, so deleting the row would orphan encrypted-and-unrecoverable content (the
-///    wrapped key lives on the row we'd delete). Tell the user to unlock first.
-///  - **LOCKED + SESSION-UNLOCKED →** PERMANENTLY remove the lock first (`remove_lock_inner`:
-///    KEK → unwrap CK → decrypt every note/transcript/timeline/audio back to plaintext, re-export the
-///    `.md`, clear the blobs, flip the folder open). Only then does it become the OPEN case below, so
-///    nothing is ever left encrypted-and-orphaned.
-///  - **OPEN (now) →** move every note to the vault ROOT (`folder_id = NULL`), delete the folder row,
-///    and remove the (now-empty) vault subdir. Notes survive at "All notes".
-#[tauri::command]
-pub fn delete_folder(state: State<'_, AppState>, folder_id: String) -> Result<(), AppError> {
-    delete_folder_inner(state.inner(), folder_id)
-}
-
-/// Inner of [`delete_folder`] taking `&AppState` (so tests can drive it without a `tauri::State`).
-/// See the command doc for the fail-closed rules.
-pub(crate) fn delete_folder_inner(state: &AppState, folder_id: String) -> Result<(), AppError> {
-    let folder = state
-        .db
-        .folder_by_id(&folder_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no folder {folder_id}")))?;
-
-    // Refuse a non-empty SUBTREE — never orphan child folders by dangling their parent_id.
-    if !state.db.child_folders(&folder_id)?.is_empty() {
-        return Err(AppError::InvalidArg(
-            "this folder has subfolders — delete or move them first".into(),
-        ));
-    }
-
-    // If sealed, it MUST be session-unlocked so we can unseal its notes back to plaintext before the
-    // folder row (which carries the wrapped key) is destroyed. Otherwise refuse — never orphan
-    // sealed content.
-    if folder.locked {
-        let session_unlocked = state
-            .unlocked_folders
-            .lock()
-            .map_err(|_| AppError::Storage("unlocked-folders mutex poisoned".into()))?
-            .contains(&folder_id);
-        if !session_unlocked {
-            return Err(AppError::Locked(
-                "unlock this folder first to delete it (its notes are sealed)".into(),
-            ));
-        }
-        // Permanently unseal back to plaintext + re-export the `.md`s, then the folder is OPEN.
-        // remove_lock_inner takes the lifecycle guard itself (so we do NOT hold it across this call —
-        // the std Mutex is non-reentrant and would self-deadlock).
-        remove_lock_inner(state, folder_id.clone())?;
-    }
-
-    // OPEN folder now (or was open all along): move its notes to the vault ROOT, then drop the row.
-    // Serialize the reassign + row delete + FS cleanup under the lifecycle guard so it can't race a
-    // concurrent lock/move on the same folder.
-    let _lifecycle = lifecycle_guard(state);
-
-    // Move every note in this folder to the vault root (folder_id = NULL). The notes' plaintext `.md`
-    // files already live in this folder's vault subdir; we re-point each meeting's exported_path to
-    // the root by moving the file (best-effort, copy-then-remove — never loses bytes).
-    let notes = state.db.notes_in_folder(&folder_id)?;
-    let mut moved_meetings = std::collections::HashSet::new();
-    for n in &notes {
-        if !moved_meetings.insert(n.meeting_id.clone()) {
-            continue;
-        }
-        // Reassign every provider row of this meeting to the root.
-        state.db.set_meeting_folder(&n.meeting_id, None)?;
-        // Best-effort move of the plaintext `.md` to the vault root (only when one exists).
-        if let Some(src_path) = n.exported_path.clone() {
-            if let Some(vault) = vault_path(state) {
-                move_note_file_to_root(state, &n.meeting_id, &src_path, &vault)?;
-            }
-        }
-    }
-
-    // NOTES-1 (2026-07-11 audit, CRITICAL data loss): AUTHORED notes (`documents(kind='note')`)
-    // must be REPARENTED to the default note-folder, NOT destroyed — the FE promises "delete folder"
-    // MOVES its notes to the default folder. The pre-fix path left them for `Db::delete_folder`'s
-    // blanket `DELETE FROM documents`, permanently deleting authored notes. `Db::delete_folder` now
-    // REFUSES if any authored note still references the folder, so this reparent MUST run first.
-    reparent_authored_notes_to_default(state, &folder_id)?;
-
-    // Delete the folder row, then remove the (now note-free) vault subdir. Row first: a leftover
-    // empty dir is harmless/reconcilable; a dangling row is not.
-    state.db.delete_folder(&folder_id)?;
-    if let Some(vault) = vault_path(state) {
-        let vault_root = std::path::Path::new(&vault);
-        if let Ok(dir) = assert_in_vault(vault_root, std::path::Path::new(&folder.path)) {
-            // remove_dir (not _all): only an EMPTY dir is removed, so a stray user file is never
-            // clobbered. The notes' `.md`s were moved out above, so the dir should be empty.
-            let _ = std::fs::remove_dir(&dir);
-        }
-    }
-    Ok(())
-}
-
-/// Move a meeting's plaintext `.md` to the vault ROOT (copy-then-remove, never losing bytes) and
-/// re-point its `exported_path`. A `&AppState`-only twin of [`move_note_file`] (whose `&State`
-/// signature can't be reached from the `_inner` delete path). Used when deleting a folder demotes its
-/// notes to "All notes".
-fn move_note_file_to_root(
-    state: &AppState,
-    meeting_id: &str,
-    src_path: &str,
-    vault: &str,
-) -> Result<(), AppError> {
-    let src = std::path::Path::new(src_path);
-    let bytes = match std::fs::read_to_string(src) {
-        Ok(b) => b,
-        // Source already gone → nothing to move; the DB association is already NULL.
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-        Err(e) => return Err(AppError::Export(format!("read note for move failed: {e}"))),
-    };
-    let file_name = src
-        .file_name()
-        .and_then(|n| n.to_str())
-        .ok_or_else(|| AppError::Export("note path has no filename".into()))?;
-    let vault_root = std::path::Path::new(vault);
-    let dest = assert_in_vault(vault_root, std::path::Path::new(file_name))?;
-    let src_canon = src.canonicalize().unwrap_or_else(|_| src.to_path_buf());
-    if dest == src_canon || dest == src {
-        return Ok(()); // already at the root.
-    }
-    if let Some(dir) = dest.parent() {
-        std::fs::create_dir_all(dir)
-            .map_err(|e| AppError::Export(format!("create move dir failed: {e}")))?;
-    }
-    // Write the destination atomically, THEN remove the source (never lose bytes).
-    // Export-collision guard: bytes move verbatim → the `exported_hash` baseline is deliberately
-    // left alone (see `move_note_file` — re-stamping from moved bytes would erase the
-    // external-edit signal).
-    crate::export::overwrite_note(&dest, &bytes)?;
-    let _ = std::fs::remove_file(src);
-    if let Some(existing) = state.db.get_latest_note_for_meeting(meeting_id)? {
-        state.db.set_note_exported_path(
-            meeting_id,
-            &existing.provider_id,
-            &dest.to_string_lossy(),
-        )?;
-    }
-    Ok(())
-}
-
-/// NOTES-1 (2026-07-11 audit, CRITICAL data loss) — reparent every AUTHORED note
-/// (`documents(kind='note')`) in a to-be-deleted folder to the DEFAULT note-folder ("Notes"), moving
-/// its plaintext `.md` into the default folder's vault subdir (copy-then-remove — never lose bytes)
-/// and rewriting `documents.exported_path`. The FE's "delete folder" promises its notes MOVE to the
-/// default folder; the pre-fix path left them for `Db::delete_folder`'s blanket `DELETE`, destroying
-/// them. Runs AFTER any sealed folder was unsealed back to plaintext (`remove_lock_inner`), so the
-/// notes here are plaintext. If the folder BEING deleted IS the default note-folder itself, its notes
-/// can't reparent to themselves — REFUSE rather than risk destroying them (the FE never offers to
-/// delete the root "Notes" folder). Best-effort FS move: a note's bytes live in the DB (canonical),
-/// so a failed `.md` move never loses content — but a missing target keeps the reparent (the row moves
-/// regardless). No PII in logs (ids only).
-fn reparent_authored_notes_to_default(
-    state: &AppState,
-    folder_id: &str,
-) -> Result<(), AppError> {
-    let note_ids = state.db.note_ids_in_folder(folder_id)?;
-    if note_ids.is_empty() {
-        return Ok(());
-    }
-    let default_id = state.db.ensure_default_note_folder()?;
-    if default_id == folder_id {
-        // Deleting the default note-folder while it still holds authored notes: reparenting to
-        // itself is a no-op and the row-delete would then destroy them. Fail closed.
-        return Err(AppError::InvalidArg(
-            "cannot delete the default \"Notes\" folder while it holds notes — move them first".into(),
-        ));
-    }
-    let default_folder = state
-        .db
-        .note_folder_by_id(&default_id)?
-        .ok_or_else(|| AppError::Storage("default note-folder missing after ensure".into()))?;
-    for id in &note_ids {
-        // Reassign the row to the default note-folder (the gate/seal anchor). The default folder is
-        // OPEN (root "Notes" is never locked), so a plain reassign is correct — no reseal needed.
-        state.db.set_note_doc_folder(id, &default_id)?;
-        // Move the plaintext `.md` into the default folder's vault subdir + re-point exported_path.
-        if let Some(row) = state.db.get_note_row(id)? {
-            move_authored_note_md_to_folder(state, &row, &default_folder)?;
-        }
-    }
-    tracing::info!(
-        target: "notes",
-        folder_id = %folder_id,
-        moved = note_ids.len(),
-        "reparented authored notes to the default folder before folder delete"
-    );
-    Ok(())
-}
-
-/// Move ONE authored note's plaintext `.md` from its old export path into `target` note-folder's
-/// vault subdir (copy-then-remove — never lose bytes) and re-point `documents.exported_path`. A
-/// `&AppState`-only helper for the `_inner` delete path (which can't reach the `&State`-signature
-/// export helpers). No-op when there is no vault, no old export, or the source file is already gone
-/// (the DB row is the canonical copy; a re-export recreates the `.md`).
-fn move_authored_note_md_to_folder(
-    state: &AppState,
-    row: &crate::storage::db::NoteRow,
-    target: &NoteFolder,
-) -> Result<(), AppError> {
-    let Some(vault) = vault_path(state) else {
-        // No vault → nothing on disk to move; still clear the stale export path so a later lock
-        // never chases it.
-        state.db.set_note_doc_exported_path(&row.id, None)?;
-        return Ok(());
-    };
-    let vault_root = std::path::Path::new(&vault);
-    // Read the source bytes (if any). A missing source → nothing to move; clear the path.
-    let bytes = match row.exported_path.as_deref() {
-        Some(src_path) => match std::fs::read_to_string(src_path) {
-            Ok(b) => Some((src_path.to_string(), b)),
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => None,
-            Err(e) => return Err(AppError::Export(format!("read note for move failed: {e}"))),
-        },
-        None => None,
-    };
-    let Some((src_path, content)) = bytes else {
-        // No on-disk file → just re-export from the (canonical) DB text into the new folder if we
-        // have text; otherwise clear the stale path.
-        if row.text.is_empty() {
-            state.db.set_note_doc_exported_path(&row.id, None)?;
-        } else if let Some(p) = write_note_to_vault(state, row)? {
-            let _ = p; // write_note_to_vault already re-points exported_path.
-        }
-        return Ok(());
-    };
-    // Compose the destination inside the target folder's vault subdir, D5-contained.
-    let file_name = std::path::Path::new(&src_path)
-        .file_name()
-        .and_then(|n| n.to_str())
-        .ok_or_else(|| AppError::Export("note path has no filename".into()))?;
-    let rel = std::path::Path::new(&target.path).join(file_name);
-    let dest = assert_in_vault(vault_root, &rel)?;
-    let src_canon = std::path::Path::new(&src_path)
-        .canonicalize()
-        .unwrap_or_else(|_| std::path::PathBuf::from(&src_path));
-    if dest == src_canon {
-        // Already at the destination (nothing to move) — just record the path.
-        state
-            .db
-            .set_note_doc_exported_path(&row.id, Some(&dest.to_string_lossy()))?;
-        return Ok(());
-    }
-    if let Some(dir) = dest.parent() {
-        std::fs::create_dir_all(dir)
-            .map_err(|e| AppError::Export(format!("create move dir failed: {e}")))?;
-    }
-    // Write the destination atomically, THEN remove the source (never lose bytes).
-    // Export-collision guard: bytes move verbatim → the `exported_hash` baseline is deliberately
-    // left alone (see `move_note_file` — re-stamping from moved bytes would erase the
-    // external-edit signal).
-    crate::export::overwrite_note(&dest, &content)?;
-    let _ = std::fs::remove_file(&src_path);
-    state
-        .db
-        .set_note_doc_exported_path(&row.id, Some(&dest.to_string_lossy()))?;
-    Ok(())
 }
 
 // ── Phase 0.5 full per-folder lock: transcript + timeline + audio seal helpers ──
